requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=he9d5cce_0
altair=4.1.0=py_1
aom=3.4.0=hb486fe8_1
appnope=0.1.2=py310hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310hca72f7f_0
arrow-cpp=8.0.0=py310h72c8010_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=hb1e8313_1
aws-c-event-stream=0.1.6=h23ab428_5
aws-checksums=0.1.9=hb1e8313_0
aws-sdk-cpp=1.8.185=he271ece_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py310hecd8cb5_0
blosc=1.21.1=h97e831e_3
boost-cpp=1.79.0=h8b082ac_0
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py310hca72f7f_1002
brunsli=0.1=h23ab428_0
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
c-blosc2=2.2.0=h2699725_0
ca-certificates=2022.07.19=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h691a603_2
certifi=2022.6.15=py310hecd8cb5_0
cffi=1.15.0=py310hc55c11b_1
cfitsio=4.1.0=h2c97ad1_0
charls=2.3.4=he49afe7_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hecd8cb5_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.5=py310hecd8cb5_0
commonmark=0.9.1=pyhd3eb1b0_0
cryptography=37.0.1=py310hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py310hca72f7f_0
dask-core=2022.7.0=py310hecd8cb5_0
dataclasses=0.8=pyh6d0b6a4_7
dav1d=1.0.0=hac89ed1_1
debugpy=1.5.1=py310he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.4.2=gpl_h0bbc480_102
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
fsspec=2022.3.0=py310hecd8cb5_0
future=0.18.2=py310hecd8cb5_1
gettext=0.21.0=h7535e17_0
gflags=2.2.2=h0a44026_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.72.1=h2292cb8_0
glib-tools=2.72.1=h2292cb8_0
glog=0.5.0=h23ab428_0
gmp=6.2.1=he9d5cce_3
gnutls=3.7.6=h627f00a_2
graphite2=1.3.14=he9d5cce_1
grpc-cpp=1.46.1=h067a048_0
harfbuzz=5.1.0=h00bb2c2_0
hdf5=1.12.1=h2b2ad87_2
icu=70.1=h96cf925_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2022.7.27=py310h012e3b2_0
imageio=2.19.3=py310hecd8cb5_0
importlib-metadata=4.11.3=py310hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.1=py310hecd8cb5_0
ipython=8.4.0=py310hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jasper=2.0.33=h013e400_0
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
jsonschema=4.4.0=py310hecd8cb5_0
jupyter_client=7.2.2=py310hecd8cb5_0
jupyter_core=4.10.0=py310hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jxrlib=1.1=haf1e3a3_2
kiwisolver=1.4.2=py310he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libaec=1.0.6=he49afe7_0
libavif=0.10.1=hac89ed1_1
libblas=3.9.0=12_osx64_mkl
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcblas=3.9.0=12_osx64_mkl
libcurl=7.82.0=h6dfd666_0
libcxx=14.0.6=hce7ea42_0
libdeflate=1.12=hac89ed1_0
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h0a4fc7d_0
libffi=3.4.2=he9d5cce_4
libgfortran=5.0.0=11_2_0_h246ff09_26
libgfortran5=11.2.0=h246ff09_26
libglib=2.72.1=hfbcb929_0
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
liblapack=3.9.0=12_osx64_mkl
liblapacke=3.9.0=12_osx64_mkl
libnghttp2=1.46.0=ha29bfda_0
libopencv=4.6.0=py310hdaa94e3_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.10.0=h0a4fc7d_0
libtasn1=4.18.0=hac89ed1_1
libthrift=0.15.0=h054ceb0_0
libtiff=4.4.0=h9847915_1
libunistring=0.9.10=h9ed2024_0
libvpx=1.11.0=he9d5cce_0
libwebp=1.2.3=hf64df63_1
libwebp-base=1.2.3=hac89ed1_2
libxml2=2.9.14=h08a9926_3
libzlib=1.2.12=hfe4f2af_2
libzopfli=1.0.3=hb1e8313_0
llvm-openmp=12.0.0=h0dcd299_1
locket=1.0.0=py310hecd8cb5_0
lz4-c=1.9.3=h23ab428_1
markupsafe=2.1.1=py310hca72f7f_0
matplotlib=3.5.1=py310hecd8cb5_1
matplotlib-base=3.5.1=py310hfb0c5b7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py310hca72f7f_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
munkres=1.1.4=py_0
nbclient=0.5.13=py310hecd8cb5_0
nbconvert=6.4.4=py310hecd8cb5_0
nbformat=5.3.0=py310hecd8cb5_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=py310hecd8cb5_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py310hecd8cb5_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
notebook=6.4.11=py310hecd8cb5_0
numexpr=2.8.3=py310hdcd3fac_0
numpy=1.22.3=py310hdcd3fac_0
numpy-base=1.22.3=py310hfd2de13_0
opencv=4.6.0=py310h2ec42d9_2
openh264=2.2.0=h2292cb8_1
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1q=hca72f7f_0
orc=1.7.4=h9274d09_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310he9d5cce_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hde71d04_1
pip=22.1.2=py310hecd8cb5_0
pixman=0.40.0=h9ed2024_1
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310he9d5cce_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.6.0=py310ha188af9_2
pyarrow=8.0.0=py310h53e4f6e_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py310hecd8cb5_0
pydicom=2.3.0=pyh6c4a22f_0
pygments=2.11.2=pyhd3eb1b0_0
pympler=0.9=py_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py310hca72f7f_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.5=hdaaf3db_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.10.2=cpu_py310h30e64cd_0
pytz=2022.1=py310hecd8cb5_0
pywavelets=1.3.0=py310hca72f7f_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310he9d5cce_0
re2=2022.04.01=he9d5cce_0
readline=8.1.2=hca72f7f_1
requests=2.28.1=py310hecd8cb5_0
rich=10.16.2=pyhd3eb1b0_0
scikit-image=0.19.2=py310hc081a56_0
scikit-learn=1.1.1=py310he9d5cce_0
scipy=1.9.0=py310h70707f4_0
seaborn=0.11.2=pyhd3eb1b0_0
semver=2.13.0=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=he9d5cce_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.5=h707629a_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.11.0=py310hecd8cb5_0
svt-av1=1.1.0=hb486fe8_1
terminado=0.13.1=py310hecd8cb5_0
testpath=0.6.0=py310hecd8cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
torchvision=0.13.0=cpu_py310h87c54a7_0
tornado=6.1=py310hca72f7f_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
tzlocal=2.1=py310hecd8cb5_0
urllib3=1.26.9=py310hecd8cb5_0
utf8proc=2.6.1=h9ed2024_0
validators=0.18.2=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310hecd8cb5_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310hecd8cb5_0
x264=1!161.3030=h0d85af4_1
x265=3.5=hbb4e6a2_3
xz=5.2.5=hca72f7f_1
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zfp=0.5.5=he9d5cce_6
zipp=3.8.0=py310hecd8cb5_0
zlib=1.2.12=h4dc903c_2
zlib-ng=2.0.6=hac89ed1_0
zstd=1.5.2=hcb37349_0
